Moves Doll Exceptions
