+++ src/main/kotlin/com/hyuse/nikkeManager/domain/exceptions/doll/DollNotFoundException.kt
@@ -1,6 +1,7 @@
-package com.hyuse.nikkeManager.exceptions
+package com.hyuse.nikkeManager.domain.exceptions.doll
 
-import com.hyuse.nikkeManager.enums.Rarity
+import com.hyuse.nikkeManager.domain.enums.Rarity
+import com.hyuse.nikkeManager.domain.exceptions.nikke.NikkeException
 
 class DollNotFoundException : NikkeException {
     constructor(rarity: Rarity, level: Int) :
